Move ProcessBackupRequest to worker so that it can be called by both graphql/admin and ee package
