+++ ee/backup/backup.go
@@ -297,3 +297,4 @@ func writeKVList(list *bpb.KVList, w io.Writer) error {
 	_, err = w.Write(buf)
 	return err
 }
+
