+++ worker/backup_ee.go
@@ -14,10 +14,13 @@ package worker
 
 import (
 	"context"
+	"net/url"
+	"time"
 
 	"github.com/dgraph-io/dgraph/ee/backup"
 	"github.com/dgraph-io/dgraph/posting"
 	"github.com/dgraph-io/dgraph/protos/pb"
+	"github.com/dgraph-io/dgraph/x"
 
 	"github.com/golang/glog"
 	"github.com/pkg/errors"
@@ -69,4 +72,101 @@ func BackupGroup(ctx context.Context, in *pb.BackupRequest) (*pb.Status, error) 
 	}
 
 	return res, nil
+}
+
+func ProcessRequest(ctx context.Context, req pb.BackupRequest, forceFull bool) error {
+	if !EnterpriseEnabled() {
+		return errors.Errorf("You must enable enterprise features first. "+
+			"Supply the appropriate license file to Dgraph Zero using the HTTP endpoint.",
+			"Backup failed.")
+	}
+
+	if req.Destination == "" {
+		return errors.Errorf("You must specify a 'destination' value")
+	}
+
+	if err := x.HealthCheck(); err != nil {
+		glog.Errorf("Backup canceled, not ready to accept requests: %s", err)
+		return err
+	}
+
+	ts, err := Timestamps(ctx, &pb.Num{ReadOnly: true})
+	if err != nil {
+		glog.Errorf("Unable to retrieve readonly timestamp for backup: %s", err)
+		return err
+	}
+
+	req.ReadTs = ts.ReadOnly
+	req.UnixTs = time.Now().UTC().Format("20060102.150405.000")
+
+	// Read the manifests to get the right timestamp from which to start the backup.
+	uri, err := url.Parse(req.Destination)
+	if err != nil {
+		return err
+	}
+	handler, err := backup.NewUriHandler(uri, backup.GetCredentialsFromRequest(&req))
+	if err != nil {
+		return err
+	}
+	latestManifest, err := handler.GetLatestManifest(uri)
+	if err != nil {
+		return err
+	}
+	req.SinceTs = latestManifest.Since
+	if forceFull {
+		req.SinceTs = 0
+	}
+
+	// Update the membership state to get the latest mapping of groups to predicates.
+	if err := UpdateMembershipState(ctx); err != nil {
+		return err
+	}
+
+	// Get the current membership state and parse it for easier processing.
+	state := GetMembershipState()
+	var groups []uint32
+	predMap := make(map[uint32][]string)
+	for gid, group := range state.Groups {
+		groups = append(groups, gid)
+		predMap[gid] = make([]string, 0)
+		for pred := range group.Tablets {
+			predMap[gid] = append(predMap[gid], pred)
+		}
+	}
+
+	glog.Infof("Created backup request: %s. Groups=%v\n", &req, groups)
+	ctx, cancel := context.WithCancel(ctx)
+	defer cancel()
+
+	errCh := make(chan error, len(state.Groups))
+	for _, gid := range groups {
+		req := req
+		req.GroupId = gid
+		req.Predicates = predMap[gid]
+		go func(req *pb.BackupRequest) {
+			_, err := BackupGroup(ctx, req)
+			errCh <- err
+		}(&req)
+	}
+
+	for range groups {
+		if err := <-errCh; err != nil {
+			glog.Errorf("Error received during backup: %v", err)
+			return err
+		}
+	}
+
+	m := backup.Manifest{Since: req.ReadTs, Groups: predMap}
+	if req.SinceTs == 0 {
+		m.Type = "full"
+		m.BackupId = x.GetRandomName(1)
+		m.BackupNum = 1
+	} else {
+		m.Type = "incremental"
+		m.BackupId = latestManifest.BackupId
+		m.BackupNum = latestManifest.BackupNum + 1
+	}
+
+	bp := &backup.Processor{Request: &req}
+	return bp.CompleteBackup(ctx, &m)
 }
